% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu.R
\name{heat_units}
\alias{heat_units}
\title{Heat Units}
\usage{
heat_units(tmin, tmax, tbase = 8, topt = 30, thigh = 42)
}
\arguments{
\item{tmin}{vector of daily minimum temperatures}

\item{tmax}{vector of daily maximum temperatures}

\item{tbase}{base cardinal temperature for development}

\item{topt}{optimimum temperature for development}

\item{thigh}{maximum temperature for development}
}
\value{
vector of daily heat units
}
\description{
Calculate heat units
}
\details{
Takes a vector of daily minimum and maximum temperatures,
    and returns a vector of daily heat units
}
\author{
Matt Espe
}

