% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weather.R
\name{read_weather}
\alias{read_weather}
\title{Read Oryza Weather}
\usage{
read_weather(file_path)
}
\arguments{
\item{file_path}{The full file path for the weather file to be read in}
}
\value{
a data.frame with 9 columns (stn, year, day, srad, tmin, tmax, vp, wind, and precip).
}
\description{
Read Oryza weather files
}
\details{
Function to read weather files formated for Oryza into R.
}
\author{
Matt Espe
}

