% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_phen.R
\name{sim_phen}
\alias{sim_phen}
\title{Simulate Phenology}
\usage{
sim_phen(start, tmin, tmax, target_hu, meas_error = TRUE, ...)
}
\arguments{
\item{start}{a integer vector of start indices. The start index must
coincide with the Tmin and Tmax values. For example, if Tmin and Tmax
are provided for then entire year and the desired start date is Apr 3rd,
the day of year should be provided (94).However, if the first temperature
values are for Apr 1st, the desired start index would be 3.}

\item{tmin}{Minimum daily temperature}

\item{tmax}{Maximum daily temperature}

\item{target_hu}{The target heat units (i.e., the accumulated heat units
after which the event occurs).}

\item{meas_error}{logical, should measurement error (+/- 2 days) be added
to the results? (Note: for reproducibility, the set.seed should be used
if adding this to simulations.)}

\item{...}{Additional arguments passed to heat_units()}
}
\value{
a vector of event indexes, corresponding to the indices for
    the temperature data.
}
\description{
Simulate phenology event
}
\details{
This function simulates a phenology event given Tmin and Tmax values,
    a start (in index value of Tmin/Tmax) and a target heat unit.
    The event is considered to occur when the heat units exceed the target.
}
\author{
Matt Espe
}

