% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu.R
\name{heat_units}
\alias{heat_units}
\title{Heat Units}
\usage{
heat_units(tmin, tmax, t_base = 8, t_opt = 30, t_high = 42)
}
\arguments{
\item{tmin}{vector of daily minimum temperatures}

\item{tmax}{vector of daily maximum temperatures}

\item{t_base}{base cardinal temperature for development}

\item{t_opt}{optimimum temperature for development}

\item{t_high}{maximum temperature for development}
}
\value{
vector of daily heat units
}
\description{
Calculate heat units
}
\details{
Takes a vector of daily minimum and maximum temperatures,
    and returns a vector of daily heat units. Heat units are
    calculated by summing over approximated hourly temperatures.
}
\author{
Matt Espe
}

